these files seperates the odd numbers.
there are three files which one is test file another one is library file
and the last one is source (or) main file.
the main aim of these files are to sperate the numbers and gives user
odd numbers.
the use can give as many numbers he want in array.