this folder (remainingProblems) contains the remaining problems
in this I have created one library function file one teat file
this is the intial commit for zip file that means if the user gives two
arrays of numbers. then it will return back the one array in which
contains the array of successive numbers from two arrays.
i made this only when the two arrays are of same size