for this commit I have added one new function called findIndex
for this commit I havr changed a lot in the testv files for require I
used the destructuring concept.
and also I have added the new function which is called the find index
the usage is if the user gives you the array of numbers and will give
the a number to find the first index it should return the first index of
that number.