in this commit I have added the new function.
the function is to check whether the given array is in descending order
or not.
for that I have used a silly hack which I have taken the output from the
ascending and placed not symbol infront of the result.
this is not the correct way to do.
it will work only for some conditions.