commits for union array
for this commit I have testes for more number of possibilities and also
last time it was worked only if there are different elements in two
arrays.
but now I have changed those so, that we can give either same or
different numbers in both the array. it will work.
teste are passed